+++ metric/instrument/syncint64/syncint64.go
@@ -12,6 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package syncint64 provides synchronous instruments that accept int64
+// measurments.
+//
+// Deprecated: Use the instruments provided by
+// go.opentelemetry.io/otel/metric/instrument instead.
 package syncint64 // import "go.opentelemetry.io/otel/metric/instrument/syncint64"
 
 import (
@@ -24,6 +29,9 @@ import (
 // Counter is an instrument that records increasing values.
 //
 // Warning: methods may be added to this interface in minor releases.
+//
+// Deprecated: Use the Int64Counter in
+// go.opentelemetry.io/otel/metric/instrument instead.
 type Counter interface {
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr int64, attrs ...attribute.KeyValue)
@@ -34,6 +42,9 @@ type Counter interface {
 // UpDownCounter is an instrument that records increasing or decreasing values.
 //
 // Warning: methods may be added to this interface in minor releases.
+//
+// Deprecated: Use the Int64UpDownCounter in
+// go.opentelemetry.io/otel/metric/instrument instead.
 type UpDownCounter interface {
 	// Add records a change to the counter.
 	Add(ctx context.Context, incr int64, attrs ...attribute.KeyValue)
@@ -44,6 +55,9 @@ type UpDownCounter interface {
 // Histogram is an instrument that records a distribution of values.
 //
 // Warning: methods may be added to this interface in minor releases.
+//
+// Deprecated: Use the Int64Histogram in
+// go.opentelemetry.io/otel/metric/instrument instead.
 type Histogram interface {
 	// Record adds an additional value to the distribution.
 	Record(ctx context.Context, incr int64, attrs ...attribute.KeyValue)