+++ sdk/metric/instrument.go
@@ -19,10 +19,6 @@ import (
 
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric/instrument"
-	"go.opentelemetry.io/otel/metric/instrument/asyncfloat64"
-	"go.opentelemetry.io/otel/metric/instrument/asyncint64"
-	"go.opentelemetry.io/otel/metric/instrument/syncfloat64"
-	"go.opentelemetry.io/otel/metric/instrument/syncint64"
 	"go.opentelemetry.io/otel/metric/unit"
 	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
@@ -180,18 +176,18 @@ type instrumentImpl[N int64 | float64] struct {
 	aggregators []internal.Aggregator[N]
 }
 
-var _ asyncfloat64.Counter = &instrumentImpl[float64]{}
-var _ asyncfloat64.UpDownCounter = &instrumentImpl[float64]{}
-var _ asyncfloat64.Gauge = &instrumentImpl[float64]{}
-var _ asyncint64.Counter = &instrumentImpl[int64]{}
-var _ asyncint64.UpDownCounter = &instrumentImpl[int64]{}
-var _ asyncint64.Gauge = &instrumentImpl[int64]{}
-var _ syncfloat64.Counter = &instrumentImpl[float64]{}
-var _ syncfloat64.UpDownCounter = &instrumentImpl[float64]{}
-var _ syncfloat64.Histogram = &instrumentImpl[float64]{}
-var _ syncint64.Counter = &instrumentImpl[int64]{}
-var _ syncint64.UpDownCounter = &instrumentImpl[int64]{}
-var _ syncint64.Histogram = &instrumentImpl[int64]{}
+var _ instrument.Float64ObservableCounter = &instrumentImpl[float64]{}
+var _ instrument.Float64ObservableUpDownCounter = &instrumentImpl[float64]{}
+var _ instrument.Float64ObservableGauge = &instrumentImpl[float64]{}
+var _ instrument.Int64ObservableCounter = &instrumentImpl[int64]{}
+var _ instrument.Int64ObservableUpDownCounter = &instrumentImpl[int64]{}
+var _ instrument.Int64ObservableGauge = &instrumentImpl[int64]{}
+var _ instrument.Float64Counter = &instrumentImpl[float64]{}
+var _ instrument.Float64UpDownCounter = &instrumentImpl[float64]{}
+var _ instrument.Float64Histogram = &instrumentImpl[float64]{}
+var _ instrument.Int64Counter = &instrumentImpl[int64]{}
+var _ instrument.Int64UpDownCounter = &instrumentImpl[int64]{}
+var _ instrument.Int64Histogram = &instrumentImpl[int64]{}
 
 func (i *instrumentImpl[N]) Observe(ctx context.Context, val N, attrs ...attribute.KeyValue) {
 	// Only record a value if this is being called from the MetricProvider.